cv-industrial-defect-lab/requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.0
scikit-learn>=1.4
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9
